chore(docs): moves docs into their own package
