+++ .gitignore
@@ -8,4 +8,4 @@ node_modules
 .npmrc
 IsolatedStorage
 .docz
-docs
+/docs
