+++ packages/core/.size-limit.js
@@ -1,0 +1,26 @@
+module.exports = [
+  {
+    limit: '7 KB',
+    path: 'dist/esm/packages/core/src/index.js',
+  },
+  {
+    limit: '2.7 KB',
+    path: 'dist/esm/packages/core/src/Motion/index.js',
+  },
+  {
+    limit: '486 B',
+    path: 'dist/esm/packages/core/src/Collector/index.js',
+  },
+  {
+    limit: '129 B',
+    path: 'dist/esm/packages/core/src/FocalTarget/index.js',
+  },
+  {
+    limit: '630 B',
+    path: 'dist/esm/packages/core/src/VisibilityManager/index.js',
+  },
+  {
+    limit: '483 B',
+    path: 'dist/esm/packages/core/src/Wait/index.js',
+  },
+];
