+++ packages/docs/src/2-getting-started/images/1.png
