+++ packages/docs/src/2-getting-started/images/2.jpeg
