+++ packages/docs/src/2-getting-started/images/3.jpeg
