+++ packages/docs/src/2-getting-started/images/4.png
