+++ packages/docs/src/2-getting-started/images/5.png
