+++ packages/docs/src/2-getting-started/images/6.jpeg
