+++ packages/docs/src/2-getting-started/images/avatar.jpg
