+++ packages/docs/src/2-getting-started/images/jg.jpg
