+++ packages/docs/src/2-getting-started/styled.tsx
