+++ packages/docs/src/3-advanced-usage/docs.mdx
@@ -26,15 +26,15 @@ import {
   Divider,
 } from '@material-ui/core';
 import * as Styled from './styled';
-import { WrappedMotion as Motion } from '../../Motion';
-import { WrappedVisibilityManager as VisibilityManager } from '../../VisibilityManager';
-import CircleExpand from '../../motions/CircleExpand';
-import TripeHoverMenu from '../../motions/RevealReshapingContainer/__docz__/TripeHoverMenu';
-import FocalRevealMove from '../../motions/FocalRevealMove';
-import FocalConcealMove from '../../motions/FocalConcealMove';
-import FocalTarget from '../../FocalTarget';
-import Move from '../../motions/Move';
-import CircleShrink from '../../motions/CircleShrink';
+import { WrappedMotion as Motion } from '../../../core/src/Motion';
+import { WrappedVisibilityManager as VisibilityManager } from '../../../core/src/VisibilityManager';
+import CircleExpand from '../../../motions/src/CircleExpand';
+import TripeHoverMenu from '../../../motions/src/RevealReshapingContainer/__docz__/TripeHoverMenu';
+import FocalRevealMove from '../../../motions/src/FocalRevealMove';
+import FocalConcealMove from '../../../motions/src/FocalConcealMove';
+import FocalTarget from '../../../core/src/FocalTarget';
+import Move from '../../../motions/src/Move';
+import CircleShrink from '../../../motions/src/CircleShrink';
 import ovi from './images/ovechkin.jpg';
 import EmailChain from './EmailChain';
 import PopmotionMoveRight from './PopmotionMoveRight';
