+++ packages/docs/src/3-advanced-usage/styled.tsx
