+++ packages/docs/src/4-custom-motions/OneFullRotation.tsx
@@ -3,8 +3,8 @@ import Collector, {
   CollectorChildrenProps,
   MotionCallback,
   CollectorActions,
-} from '../../Collector';
-import { combine } from '../../lib/style';
+} from '../../../core/src/Collector';
+import { combine } from '../../../core/src/lib/style';
 
 export default class OneFullRotation extends React.Component<CollectorChildrenProps> {
   beforeAnimate: MotionCallback = (_, onFinish, setChildProps) => {
