+++ packages/docs/src/4-custom-motions/docs.mdx
@@ -6,8 +6,8 @@ order: 2
 
 import { Playground } from 'docz';
 import { Toggler } from '@element-motion/dev';
-import { WrappedMotion as Motion } from '../../Motion';
-import Move from '../../motions/Move';
+import { WrappedMotion as Motion } from '../../../core/src/Motion';
+import Move from '../../../motions/src/Move';
 import OneFullRotation from './OneFullRotation';
 import SupportingMotion from './SupportingMotion';
 import * as Styled from './styled';
